Attempt at using the router to handle persisted queries instead of DOM manipulation
