+++ packages/navigation/src/index.js
@@ -14,6 +14,8 @@ import { getHistory } from './history';
 import * as navUtils from './index';
 // For the above, import the module into itself. Functions consumed from this import can be mocked in tests.
 
+const TIME_EXCLUDED_SCREENS_FILTER = 'woocommerce_admin_time_excluded_screens';
+
 // Expose history so all uses get the same history object.
 export { getHistory };
 
@@ -44,6 +46,39 @@ export const getPersistedQuery = ( query = navUtils.getQuery() ) => {
 		'type',
 	] );
 	return pick( query, params );
+};
+
+/**
+ * Accepts a query, and returns a query that has updated persistent queries
+ * conditionally.
+ *
+ * @param {Object} query Query containing the parameters.
+ * @return {Object} Modified query object
+ */
+export const patchPersistedQueries = ( query ) => {
+	const excludedScreens = applyFilters( TIME_EXCLUDED_SCREENS_FILTER, [
+		'stock',
+		'settings',
+		'customers',
+		'homescreen',
+	] );
+
+	const path = query.path || 'homescreen';
+	const screen = path.replace( '/analytics', '' ).replace( '/', '' );
+
+	if ( excludedScreens.includes( screen ) ) {
+		return query;
+	}
+
+	const persisted = getPersistedQuery();
+	const patchedQuery =
+		query.updatePersistedQueries && query.updatePersistedQueries === 'true'
+			? Object.assign( persisted, query )
+			: Object.assign( query, persisted );
+
+	delete patchedQuery.updatePersistedQueries;
+
+	return patchedQuery;
 };
 
 /**
